Added MessageBox to confirm when deleting a group

+++ Views/Commands/DeleteGroupCommand.cs
@@ -3,9 +3,11 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
 using System.Windows.Input;
 using Whydoisuck.DataSaving;
 using Whydoisuck.Model.DataStructures;
+using Whydoisuck.Properties;
 
 namespace Whydoisuck.Views.Commands
 {
@@ -29,7 +31,14 @@ namespace Whydoisuck.Views.Commands
 
         public void Execute(object parameter)
         {
-            SessionManager.Instance.DeleteGroup(Group);
+            var caption = string.Format(Resources.DeleteGroupCaptionFormat, Group.GroupName);
+            var content = string.Format(Resources.DeleteGroupContentFormat, Group.GroupName);
+            // lazyness, TODO custom dialog
+            var result = MessageBox.Show(content, caption, MessageBoxButton.YesNo, MessageBoxImage.Warning, MessageBoxResult.No);
+            if(result == MessageBoxResult.Yes)
+            {
+                SessionManager.Instance.DeleteGroup(Group);
+            }
         }
     }
 }
